requirements_all.txt
### Full Environment (Inference + Download + MPC)
affine==2.4.0
appdirs==1.4.4
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.9.0
fonttools==4.45.1
fsspec==2023.4.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
nvidia-ml-py3==7.352.0
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
portalocker==2.8.2
protobuf==4.25.1
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rasterio==1.3.9
requests==2.28.1
scipy==1.11.4
sentry-sdk==1.37.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
sympy==1.13.1
tabulate==0.9.0
termcolor==2.3.0 
tqdm==4.66.1
triton==3.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.16.0
yacs==0.1.8
fiona
geopandas
odc-stac
pystac-client
planetary-computer
stackstac
rioxarray
dask
earthengine-api
tqdm
